pyproject.toml
[build-system]
requires = [ "setuptools>=68.0" ]
build-backend = "setuptools.build_meta"

[project]
name = "ddeutil-extensions"
description = "Extension functions and objects"
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
authors = [{ name = "ddeutils", email = "korawich.anu@gmail.com" }]
keywords = ['extensions', 'utility']
classifiers = [
    "Topic :: Utilities",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9.13"
dependencies = [
    "fmtutil==1.0.15",
    "ddeutil-workflow==0.0.61"
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/ddeutils/ddeutil-extensions/"
"Source Code" = "https://github.com/ddeutils/ddeutil-extensions/"

[project.optional-dependencies]
test = [
    "python-dotenv==1.1.0",
    "sqlalchemy==2.0.40",
    "fsspec==2025.3.2",
    "ddeutil-extensions[dataframe,sftp,aws,deltalake]",
]
dataframe = [
    "ddeutil-extensions[polars,pandas,duckdb]",
]
polars = [
    "polars==1.29.0",
    # NOTE: Read Excel file.
    "fastexcel==0.14.0",
]
pandas = [
    "pandas==2.2.3",
]
sftp = [
    "paramiko==3.5.1",
    "sshtunnel==0.4.0",
]
aws = [
    "boto3==1.38.6",
]
deltalake = [
    "deltalake==0.25.5",
]
duckdb = [
    "duckdb==1.2.2",
]
restapi = [
    "requests==2.32.3",
    "httpx==0.28.1",
]

[tool.setuptools.dynamic]
version = {attr = "ddeutil.extensions.__about__.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.shelf.version]
version = "./src/ddeutil/extensions/__about__.py"
changelog = "CHANGELOG.md"
commit_msg_format = "- {subject}"

[tool.shelf.git]
commit_prefix_force_fix = true

[tool.coverage.run]
branch = true
relative_files = true
concurrency = ["thread", "multiprocessing"]
source = ["ddeutil", "tests"]

[tool.coverage.report]
exclude_lines = ["raise NotImplementedError"]

[tool.pytest.ini_options]
pythonpath = ["src"]
console_output_style = "count"
addopts = [
    "--strict-config",
    "--strict-markers",
]
filterwarnings = ["error"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)7s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.black]
line-length = 80
target-version = ['py39']
exclude = """
(
    /(
        \\.git
        | \\.eggs
        | \\.__pycache__
        | \\.idea
        | \\.ruff_cache
        | \\.mypy_cache
        | \\.pytest_cache
        | \\.venv
        | build
        | dist
        | venv
    )/
)
"""

[tool.ruff]
line-length = 80
exclude = [
    "__pypackages__",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",
    "F403",
]
